@@ -1,22 +1,32 @@
+/**
+ * @file
+ * Custom JS for the Backup and Migrate module.
+ */
+
 (function($) {
+  'use strict';
+
   Drupal.behaviors.backupMigrate = {
-    attach: function(context) {
-      if (Drupal.settings.backup_migrate !== undefined) { 
+    attach: function(context, settings) {
+      if (Drupal.settings.backup_migrate !== undefined) {
         if (Drupal.settings.backup_migrate.dependents !== undefined) {
+          var key;
           for (key in Drupal.settings.backup_migrate.dependents) {
             info = Drupal.settings.backup_migrate.dependents[key];
-            dependent = $('#edit-' + info['dependent']);
+            var dependent = $('#edit-' + info['dependent']);
             for (key in info['dependencies']) {
               $('[name="' + key + '"]').each(function() {
                 var dependentval = info['dependencies'][key];
                 var dependency = $(this);
                 (function(dependent, dependency) {
                   var checkval = function(inval) {
-                    if (dependency.attr('type') == 'radio') {
+                    // Do loose comparisons to support things like "true", "1",
+                    // etc.
+                    if (dependency.attr('type') === 'radio') {
                       var val = $('[name="' + dependency.attr('name') + '"]:checked').val();
                       return val == inval;
                     }
-                    else if (dependency.attr('type') == 'checkbox') {
+                    else if (dependency.attr('type') === 'checkbox') {
                       return dependency.is(':checked') && inval == dependency.val();
                     }
                     else {
@@ -53,12 +63,13 @@
                 var label = $(this).attr('label');
                 copy_selector_options[label] = [];
                 $(this).find('option').each(function() {
-                  copy_selector_options[label].push(this); 
+                  copy_selector_options[label].push(this);
                 });
                 $(this).remove();
               })
 
-              // Assign an action to the main selector to modify the secondary selector
+              // Assign an action to the main selector to modify the secondary
+              // selector.
               selector.each(function() {
                 $(this).bind('load change click keypress focus', function() {
                   var group = $(this).find('option[value=' + $(this).val() + ']').parents('optgroup').attr('label');
@@ -75,27 +86,28 @@
               });
             })(info);
           }
+
           // Add the convert to checkboxes functionality to all multiselects.
           $('#backup-migrate-ui-manual-backup-form select[multiple], #backup-migrate-crud-edit-form select[multiple]').each(function() {
             var self = this;
             $(self).after(
               $('<div class="description backup-migrate-checkbox-link"></div>').append(
-                $('<a>'+ Drupal.settings.backup_migrate.checkboxLinkText +'</a>').click(function() {
+                $('<a>' + Drupal.settings.backup_migrate.checkboxLinkText + '</a>').click(function() {
                   var $select = $(self);
                   var $checkboxes = $('<div></div>').addClass('backup-migrate-tables-checkboxes');
                   $('option', $select).each(function(i) {
                     $checkboxes.append(
                       $('<div class="form-item"></div>').append(
                         $('<label class="option backup-migrate-table-select">' + this.value + '</label>').prepend(
-                          $('<input type="checkbox" class="backup-migrate-tables-checkbox" name="'+ $select.attr('name') +'"'+ (this.selected ? 'checked="checked"' : '') +' value="'+ this.value +'"/>')
+                          $('<input type="checkbox" class="backup-migrate-tables-checkbox" name="' + $select.attr('name') + '"' + (this.selected ? 'checked="checked"' : '') + ' value="' + this.value + '"/>')
                             .bind('click change load', function() {
-                                if (this.checked) {
-                                  $(this).parent().addClass('checked');
-                                }
-                                else {
-                                  $(this).parent().removeClass('checked');
-                                }
-                              }).load()
+                              if (this.checked) {
+                                $(this).parent().addClass('checked');
+                              }
+                              else {
+                                $(this).parent().removeClass('checked');
+                              }
+                            }).load()
                         )
                       )
                     );
