@@ -3,7 +3,7 @@
  * This is a PHP library that handles calling reCAPTCHA.
  *
  * @copyright Copyright (c) 2015, Google Inc.
- * @link      http://www.google.com/recaptcha
+ * @link      https://www.google.com/recaptcha
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -26,6 +26,7 @@
 
 namespace ReCaptcha\RequestMethod;
 
+use ReCaptcha\ReCaptcha;
 use ReCaptcha\RequestMethod;
 use ReCaptcha\RequestParameters;
 
@@ -37,24 +38,27 @@ use ReCaptcha\RequestParameters;
 class CurlPost implements RequestMethod
 {
     /**
-     * URL to which requests are sent via cURL.
-     * @const string
-     */
-    const SITE_VERIFY_URL = 'https://www.google.com/recaptcha/api/siteverify';
-
-    /**
      * Curl connection to the reCAPTCHA service
      * @var Curl
      */
     private $curl;
 
-    public function __construct(Curl $curl = null)
+    /**
+     * URL for reCAPTCHA sitevrerify API
+     * @var string
+     */
+    private $siteVerifyUrl;
+
+    /**
+     * Only needed if you want to override the defaults
+     *
+     * @param Curl $curl Curl resource
+     * @param string $siteVerifyUrl URL for reCAPTCHA sitevrerify API
+     */
+    public function __construct(Curl $curl = null, $siteVerifyUrl = null)
     {
-        if (!is_null($curl)) {
-            $this->curl = $curl;
-        } else {
-            $this->curl = new Curl();
-        }
+        $this->curl = (is_null($curl)) ? new Curl() : $curl;
+        $this->siteVerifyUrl = (is_null($siteVerifyUrl)) ? ReCaptcha::SITE_VERIFY_URL : $siteVerifyUrl;
     }
 
     /**
@@ -65,7 +69,7 @@ class CurlPost implements RequestMethod
      */
     public function submit(RequestParameters $params)
     {
-        $handle = $this->curl->init(self::SITE_VERIFY_URL);
+        $handle = $this->curl->init($this->siteVerifyUrl);
 
         $options = array(
             CURLOPT_POST => true,
@@ -83,6 +87,10 @@ class CurlPost implements RequestMethod
         $response = $this->curl->exec($handle);
         $this->curl->close($handle);
 
-        return $response;
+        if ($response !== false) {
+            return $response;
+        }
+
+        return '{"success": false, "error-codes": ["'.ReCaptcha::E_CONNECTION_FAILED.'"]}';
     }
 }
