@@ -3,7 +3,7 @@
  * This is a PHP library that handles calling reCAPTCHA.
  *
  * @copyright Copyright (c) 2015, Google Inc.
- * @link      http://www.google.com/recaptcha
+ * @link      https://www.google.com/recaptcha
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -36,7 +36,7 @@ class Socket
 
     /**
      * fsockopen
-     * 
+     *
      * @see http://php.net/fsockopen
      * @param string $hostname
      * @param int $port
@@ -51,14 +51,13 @@ class Socket
 
         if ($this->handle != false && $errno === 0 && $errstr === '') {
             return $this->handle;
-        } else {
-            return false;
         }
+        return false;
     }
 
     /**
      * fwrite
-     * 
+     *
      * @see http://php.net/fwrite
      * @param string $string
      * @param int $length
@@ -71,7 +70,7 @@ class Socket
 
     /**
      * fgets
-     * 
+     *
      * @see http://php.net/fgets
      * @param int $length
      * @return string
@@ -83,7 +82,7 @@ class Socket
 
     /**
      * feof
-     * 
+     *
      * @see http://php.net/feof
      * @return bool
      */
@@ -94,7 +93,7 @@ class Socket
 
     /**
      * fclose
-     * 
+     *
      * @see http://php.net/fclose
      * @return bool
      */
