@@ -3,7 +3,7 @@
  * This is a PHP library that handles calling reCAPTCHA.
  *
  * @copyright Copyright (c) 2015, Google Inc.
- * @link      http://www.google.com/recaptcha
+ * @link      https://www.google.com/recaptcha
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -26,37 +26,17 @@
 
 namespace ReCaptcha\RequestMethod;
 
+use ReCaptcha\ReCaptcha;
 use ReCaptcha\RequestMethod;
 use ReCaptcha\RequestParameters;
 
 /**
  * Sends a POST request to the reCAPTCHA service, but makes use of fsockopen()
  * instead of get_file_contents(). This is to account for people who may be on
- * servers where allow_furl_open is disabled.
+ * servers where allow_url_open is disabled.
  */
 class SocketPost implements RequestMethod
 {
-    /**
-     * reCAPTCHA service host.
-     * @const string
-     */
-    const RECAPTCHA_HOST = 'www.google.com';
-
-    /**
-     * @const string reCAPTCHA service path
-     */
-    const SITE_VERIFY_PATH = '/recaptcha/api/siteverify';
-
-    /**
-     * @const string Bad request error
-     */
-    const BAD_REQUEST = '{"success": false, "error-codes": ["invalid-request"]}';
-
-    /**
-     * @const string Bad response error
-     */
-    const BAD_RESPONSE = '{"success": false, "error-codes": ["invalid-response"]}';
-
     /**
      * Socket to the reCAPTCHA service
      * @var Socket
@@ -64,17 +44,15 @@ class SocketPost implements RequestMethod
     private $socket;
 
     /**
-     * Constructor
+     * Only needed if you want to override the defaults
      *
      * @param \ReCaptcha\RequestMethod\Socket $socket optional socket, injectable for testing
+     * @param string $siteVerifyUrl URL for reCAPTCHA sitevrerify API
      */
-    public function __construct(Socket $socket = null)
+    public function __construct(Socket $socket = null, $siteVerifyUrl = null)
     {
-        if (!is_null($socket)) {
-            $this->socket = $socket;
-        } else {
-            $this->socket = new Socket();
-        }
+        $this->socket = (is_null($socket)) ? new Socket() : $socket;
+        $this->siteVerifyUrl = (is_null($siteVerifyUrl)) ? ReCaptcha::SITE_VERIFY_URL : $siteVerifyUrl;
     }
 
     /**
@@ -87,15 +65,16 @@ class SocketPost implements RequestMethod
     {
         $errno = 0;
         $errstr = '';
+        $urlParsed = parse_url($this->siteVerifyUrl);
 
-        if (false === $this->socket->fsockopen('ssl://' . self::RECAPTCHA_HOST, 443, $errno, $errstr, 30)) {
-            return self::BAD_REQUEST;
+        if (false === $this->socket->fsockopen('ssl://' . $urlParsed['host'], 443, $errno, $errstr, 30)) {
+            return '{"success": false, "error-codes": ["'.ReCaptcha::E_CONNECTION_FAILED.'"]}';
         }
 
         $content = $params->toQueryString();
 
-        $request = "POST " . self::SITE_VERIFY_PATH . " HTTP/1.1\r\n";
-        $request .= "Host: " . self::RECAPTCHA_HOST . "\r\n";
+        $request = "POST " . $urlParsed['path'] . " HTTP/1.1\r\n";
+        $request .= "Host: " . $urlParsed['host'] . "\r\n";
         $request .= "Content-Type: application/x-www-form-urlencoded\r\n";
         $request .= "Content-length: " . strlen($content) . "\r\n";
         $request .= "Connection: close\r\n\r\n";
@@ -111,7 +90,7 @@ class SocketPost implements RequestMethod
         $this->socket->fclose();
 
         if (0 !== strpos($response, 'HTTP/1.1 200 OK')) {
-            return self::BAD_RESPONSE;
+            return '{"success": false, "error-codes": ["'.ReCaptcha::E_BAD_RESPONSE.'"]}';
         }
 
         $parts = preg_split("#\n\s*\n#Uis", $response);
