@@ -3,7 +3,7 @@
  * This is a PHP library that handles calling reCAPTCHA.
  *
  * @copyright Copyright (c) 2015, Google Inc.
- * @link      http://www.google.com/recaptcha
+ * @link      https://www.google.com/recaptcha
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -26,14 +26,18 @@
 
 namespace ReCaptcha\RequestMethod;
 
+use ReCaptcha\ReCaptcha;
 use ReCaptcha\RequestParameters;
+use PHPUnit\Framework\TestCase;
 
-class SocketPostTest extends \PHPUnit_Framework_TestCase
+class SocketPostTest extends TestCase
 {
-
     public function testSubmitSuccess()
     {
-        $socket = $this->getMock('\\ReCaptcha\\RequestMethod\\Socket', array('fsockopen', 'fwrite', 'fgets', 'feof', 'fclose'));
+        $socket = $this->getMockBuilder(\ReCaptcha\RequestMethod\Socket::class)
+            ->disableOriginalConstructor()
+            ->setMethods(array('fsockopen', 'fwrite', 'fgets', 'feof', 'fclose'))
+            ->getMock();
         $socket->expects($this->once())
                 ->method('fsockopen')
                 ->willReturn(true);
@@ -54,9 +58,40 @@ class SocketPostTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('RESPONSEBODY', $response);
     }
 
+    public function testOverrideSiteVerifyUrl()
+    {
+        $socket = $this->getMockBuilder(\ReCaptcha\RequestMethod\Socket::class)
+            ->disableOriginalConstructor()
+            ->setMethods(array('fsockopen', 'fwrite', 'fgets', 'feof', 'fclose'))
+            ->getMock();
+        $socket->expects($this->once())
+                ->method('fsockopen')
+                ->with('ssl://over.ride', 443, 0, '', 30)
+                ->willReturn(true);
+        $socket->expects($this->once())
+                ->method('fwrite')
+                ->with($this->matchesRegularExpression('/^POST \/some\/path.*Host: over\.ride/s'));
+        $socket->expects($this->once())
+                ->method('fgets')
+                ->willReturn("HTTP/1.1 200 OK\n\nRESPONSEBODY");
+        $socket->expects($this->exactly(2))
+                ->method('feof')
+                ->will($this->onConsecutiveCalls(false, true));
+        $socket->expects($this->once())
+                ->method('fclose')
+                ->willReturn(true);
+
+        $ps = new SocketPost($socket, 'https://over.ride/some/path');
+        $response = $ps->submit(new RequestParameters("secret", "response", "remoteip", "version"));
+        $this->assertEquals('RESPONSEBODY', $response);
+    }
+
     public function testSubmitBadResponse()
     {
-        $socket = $this->getMock('\\ReCaptcha\\RequestMethod\\Socket', array('fsockopen', 'fwrite', 'fgets', 'feof', 'fclose'));
+        $socket = $this->getMockBuilder(\ReCaptcha\RequestMethod\Socket::class)
+            ->disableOriginalConstructor()
+            ->setMethods(array('fsockopen', 'fwrite', 'fgets', 'feof', 'fclose'))
+            ->getMock();
         $socket->expects($this->once())
                 ->method('fsockopen')
                 ->willReturn(true);
@@ -74,17 +109,20 @@ class SocketPostTest extends \PHPUnit_Framework_TestCase
 
         $ps = new SocketPost($socket);
         $response = $ps->submit(new RequestParameters("secret", "response", "remoteip", "version"));
-        $this->assertEquals(SocketPost::BAD_RESPONSE, $response);
+        $this->assertEquals('{"success": false, "error-codes": ["'.ReCaptcha::E_BAD_RESPONSE.'"]}', $response);
     }
 
-    public function testSubmitBadRequest()
+    public function testConnectionFailureReturnsError()
     {
-        $socket = $this->getMock('\\ReCaptcha\\RequestMethod\\Socket', array('fsockopen'));
+        $socket = $this->getMockBuilder(\ReCaptcha\RequestMethod\Socket::class)
+            ->disableOriginalConstructor()
+            ->setMethods(array('fsockopen'))
+            ->getMock();
         $socket->expects($this->once())
                 ->method('fsockopen')
                 ->willReturn(false);
         $ps = new SocketPost($socket);
         $response = $ps->submit(new RequestParameters("secret", "response", "remoteip", "version"));
-        $this->assertEquals(SocketPost::BAD_REQUEST, $response);
+        $this->assertEquals('{"success": false, "error-codes": ["'.ReCaptcha::E_CONNECTION_FAILED.'"]}', $response);
     }
 }
