@@ -7,6 +7,7 @@
 
 namespace ReCaptcha\RequestMethod;
 
+use ReCaptcha\ReCaptcha;
 use ReCaptcha\RequestMethod;
 use ReCaptcha\RequestParameters;
 
@@ -16,16 +17,13 @@ use ReCaptcha\RequestParameters;
 class Drupal7Post implements RequestMethod {
 
   /**
-   * URL to which requests are POSTed.
-   * @const string
-   */
-  const SITE_VERIFY_URL = 'https://www.google.com/recaptcha/api/siteverify';
-
-  /**
    * Submit the POST request with the specified parameters.
    *
-   * @param RequestParameters $params Request parameters
-   * @return string Body of the reCAPTCHA response
+   * @param ReCaptcha\RequestParameters $params
+   *   Request parameters.
+   *
+   * @return string
+   *   Body of the reCAPTCHA response.
    */
   public function submit(RequestParameters $params) {
 
@@ -36,8 +34,20 @@ class Drupal7Post implements RequestMethod {
       'method' => 'POST',
       'data' => $params->toQueryString(),
     );
-    $response = drupal_http_request(self::SITE_VERIFY_URL, $options);
+    $response = drupal_http_request(ReCaptcha::SITE_VERIFY_URL, $options);
 
-    return isset($response->data) ? $response->data : '';
+    if ($response->code == 200 && isset($response->data)) {
+      // The service request was successful.
+      return $response->data;
+    }
+    elseif ($response->code < 0) {
+      // Negative status codes typically point to network or socket issues.
+      return '{"success": false, "error-codes": ["' . ReCaptcha::E_CONNECTION_FAILED . '"]}';
+    }
+    else {
+      // Positive none 200 status code typically means the request has failed.
+      return '{"success": false, "error-codes": ["' . ReCaptcha::E_BAD_RESPONSE . '"]}';
+    }
   }
+
 }
